pyproject.toml
[tool.poetry]
name = "fastapi-hotels"
version = "0.1.0"
description = "fastapi stepik course project"
authors = ["NelttjeN <tatyapa0880@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.0"}
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
alembic = "^1.11.3"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-json-logger = "^2.0.7"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.4"
pillow = "^10.0.0"
pymongo = "^4.5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
flake8-builtins = "^2.1.0"
flake8-commas = "^2.1.0"
flake8-functions-names = "^0.4.0"
flake8-import-order = "^0.18.2"
flake8-length = "^0.3.1"
flake8-quotes = "^3.3.2"
flake8-variables-names = "^0.0.6"
flake8_simplify = "^0.20.0"
pep8-naming = "^0.13.3"
coverage = "^7.3.0"
pyflakes = "^3.1.0"
pipdeptree = "^2.13.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pip-shims = "0.3.4"
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
